Add reusable thread pool
This thread pool is designed for the use case where there are many
batches of similar tasks. The thread pool can be waited on without
requiring the creation of threads again, and in each batch resources can
be shared between tasks that happen to be done on a single thread.
For example, we may only allocate a buffer once per batch per thread, or
even fewer times if tasks in separated batches are the same, instead of
allocating it for every single task.